Encoder reading is good now. Have done an attempt on PID, but not tested yet.

+++ Node1/CAN.h
@@ -17,9 +17,9 @@
 /** Struct for message, defining id, length and data byte.
  */
 typedef struct {
-    unsigned int id;
+    uint8_t id;
     uint8_t length;
-    int8_t data[8]; 
+    uint8_t data[8]; 
 } message;
